Add in checks for jones_array. Add axis descriptions to second table.
added checks for jones array values to be contiguous. Thow ValueError if not.

Add in axis descritions (CTYPE's) for the secondary table which holds flag_array.
This second table is only used if the calibration type is delay. Otherwise,
the flags are in the primary data array. Updated both read and write functions
to reflect this change.