Use smaller but sufficient data types to reduce memory usage.
MWA data comes in 32-bit floats; when combining each pair of float to make a
complex number, the appropriate type is np.complex64, not np.complex128.

I don't know the details behind UVData's nsample_array, but at least in the MWA
context, this is only ever set to 1 or 0. It would be better to use bools or
8-bit ints, but pyuvdata insists that this type should be a float, and
np.float16 is the smallest.

I have verified with a run of SSINS over MWA data that the output is the same,
and the memory usage reduced from ~78 GB to ~44 GB.