Update uvdata_tutorial.rst
Updated plt.imshow(np.abs(uvd.data_array[bl_ind, 0, :, 0])) to instead be plt.imshow(np.abs(uvd.data_array[bl_ind, :, 0])), to take into account the use of the use_future_array_shapes=True boolean when using the UVData.from_file() method in this chunk of the tutorial. This boolean is described in documentation as: "Option to convert to the future planned array shapes before the changes go into effect by removing the spectral window axis."

On a cursory glance, this appears to be the only place in the tutorial where this comes up, but it could be worth checking that this is in fact the case.